Added Chapter 2 - Section 2.3
This section goes into detail about Rooms and Regions and why you should use them. You are taught how to create a Dark Room class that inherits from the Room class.  We discuss the "cannotGoThatWayMsg" Property, and a few excersise is given to the reader for better practice of developing more complex maps.